REFACTOR: Move and rename a couple of function in thetawave_web::public_s3_assets
